requirements.txt
streamlit==1.11.0
finance-datareader
beautifulsoup4
streamlit_lottie==0.0.2
requests==2.24.0
Pillow==8.4.0
protobuf~=3.19.0
watchdog==2.1.8
